+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Log failed PostgreSQL version check with `ERROR` level.
 - Provide better error message when available space can't be determined.
+- Set `me` parameter of emoji reaction to `true` if it was made by current user.
 
 ## [3.1.0] - 2024-08-08